refactor(handler): 纯移动拆分 setting_handler.go（3957→468行）
